# specs/001-ai-resource-hub/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Resource Hub (Feature Contract)
  version: 0.1.0
  description: |
    Contract additions for feature 001-ai-resource-hub.

    Note: This is a feature-level contract intended to guide implementation.
    The repo's canonical OpenAPI remains the running backend at /api/v1/openapi.json.
servers:
  - url: /api/v1
...
paths:
...
  /submissions:
    get:
      tags: [submissions]
      summary: List submissions visible to signed-in users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: state
          schema: { type: string, enum: [pending, approved, rejected] }
        - in: query
          name: skip
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionsPublic'
    post:
      tags: [submissions]
      summary: Create submission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPublic'
        '409':
          description: Conflict (destination_url already exists as a Resource)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /submissions/{id}:
    get:
      tags: [submissions]
      summary: Get submission by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPublic'
    put:
      tags: [submissions]
      summary: Update my submission (pending only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPublic'
  /submissions/{id}/comments:
    get:
      tags: [submission-comments]
      summary: List comments for a submission
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsPublic'
    post:
      tags: [submission-comments]
      summary: Create comment on a submission
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPublic'
  /admin/submissions/{id}/approve:
    post:
      tags: [admin]
      summary: Approve a submission
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePublic'
        '409':
          description: Conflict (destination_url already exists as a Resource)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /admin/submissions/{id}/reject:
    post:
      tags: [admin]
      summary: Reject a submission
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPublic'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        message: { type: string }
      required: [message]
    ResourceType:
      type: string
      enum: [github_repo, tool, website, article, other]
    ResourceBase:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { $ref: '#/components/schemas/ResourceType' }
        destination_url: { type: string, format: uri }
      required: [title, description, type, destination_url]
...
    ResourcePublic:
      allOf:
        - $ref: '#/components/schemas/ResourceBase'
        - type: object
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            like_count: { type: integer }
            favorite_count: { type: integer }
            comment_count: { type: integer }
          required: [id, created_at, updated_at]
...
    SubmissionState:
      type: string
      enum: [pending, approved, rejected]
    SubmissionBase:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { $ref: '#/components/schemas/ResourceType' }
        destination_url: { type: string, format: uri }
      required: [title, description, type, destination_url]
    SubmissionCreate:
      allOf:
        - $ref: '#/components/schemas/SubmissionBase'
    SubmissionUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { $ref: '#/components/schemas/ResourceType' }
        destination_url: { type: string, format: uri }
    SubmissionPublic:
      allOf:
        - $ref: '#/components/schemas/SubmissionBase'
        - type: object
          properties:
            id: { type: string, format: uuid }
            state: { $ref: '#/components/schemas/SubmissionState' }
            submitter_id: { type: string, format: uuid }
            reviewed_by_id: { type: string, format: uuid, nullable: true }
            reviewed_at: { type: string, format: date-time, nullable: true }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
          required: [id, state, submitter_id, created_at, updated_at]
    SubmissionsPublic:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SubmissionPublic' }
        count: { type: integer }
      required: [data, count]
    CommentCreate:
      type: object
      properties:
        body: { type: string }
      required: [body]
...
    CommentPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        author_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, body, author_id, created_at, updated_at]
    CommentsPublic:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/CommentPublic' }
        count: { type: integer }
      required: [data, count]
...